+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+V1.0.0
+ - Removed legacy plugins 'any', 'all', 'each', 'order_by', 'select_attr', 'select_many', 'where', 'where_compare', 'delay', 'assert_function'
+ - Added is_unknown plugin
+
 V0.26.0
  - Removed in-band signaling for files
  - Removed snapshot restore functionality